Show resource stats for limits too
In reality, resource utilization can be higher than the sum of the
requests of pods in the cluster. Especially for cluster operators who
want node stability, resource limits are much more relevant than
resource requests.

I tried to keep the output format relatively similar as before, though I
thought it might be good to be consistent with the term "CPU" instead of
"cores" and the capitalization "MEMORY" since that was used in one place
and it matches the capitalization of CPU, but that honestly matters
less than just adding output for limits.